add cool/warm tooltip, move gustavo's tooltip code
